requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
greenlet==2.0.1
idna==3.4
oauthlib==3.2.2
pyben==0.3.2
Pygments==2.13.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-discogs-client==2.5
qbittorrent-api==2022.11.40
requests==2.28.1
rich==12.6.0
six==1.16.0
SQLAlchemy==1.4.44
urllib3==1.26.13
